requirements.txt
# ══════════════════════════════════════════════════
# Drizzle — Phase 2 Dependencies
# ══════════════════════════════════════════════════

# --- Core Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.0

# --- Firebase Admin SDK ---
firebase-admin>=6.4.0

# --- HTTP Client (MCP server calls) ---
httpx>=0.27.0

# --- OpenAI (Phase 1 LLM reasoning) ---
openai>=1.12.0

# --- Utilities ---
python-multipart>=0.0.9
